chore: slim test file — drop redundant cases, rename to snake_case
Removed tests that do not exercise a bug (positive equality, known-size
inequality, first-key-wins ordering). Renamed file from hyphens to
underscores to satisfy the test-name validator hook. Removed section-header
comments from the test file.
